refactor(daemon): extract runner::persistence submodule (#2642)
Move restore_state, persist_task_state, check_missed_cron_catchup,
and set_last_run to runner/persistence.rs.
